+++ src/main/java/cart/dao/product/ProductEntity.java
@@ -1,4 +1,4 @@
-package cart.dao;
+package cart.dao.product;
 
 public class ProductEntity {
