+++ src/test/java/cart/dao/ProductDaoTest.java
@@ -3,6 +3,8 @@ package cart.dao;
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
+import cart.dao.product.ProductDao;
+import cart.dao.product.ProductEntity;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
