+++ src/test/java/cart/domain/ImageTest.java
@@ -2,6 +2,7 @@ package cart.domain;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import cart.domain.product.Image;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
