+++ src/test/java/cart/domain/ProductTest.java
@@ -1,5 +1,6 @@
 package cart.domain;
 
+import cart.domain.product.Product;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
